+++ bootstrap-buildout.py
@@ -18,6 +18,7 @@ The script accepts buildout command-line options, so you can
 use the -c option to specify an alternate configuration file.
 """
 
+from __future__ import print_function
 import os
 import shutil
 import sys
@@ -73,17 +74,14 @@ parser.add_option("--setuptools-to-dir",
 
 options, args = parser.parse_args()
 if options.version:
-    print("bootstrap.py version %s" % __version__)
+    print(("bootstrap.py version %s" % __version__))
     sys.exit(0)
 
 
 ######################################################################
 # load/install setuptools
 
-try:
-    from urllib.request import urlopen
-except ImportError:
-    from urllib2 import urlopen
+from urllib.request import urlopen
 
 ez = {}
 if os.path.exists('ez_setup.py'):